Verbal Memory (done)/requirements.txt
Pillow
pynput
pywin32
pytesseract
pyautogui